Finished prototype anomalies.py (ref #6, #7, #8)
-) anomalies.py does calculate the anomalies as well as needed
sub-products (monthly means and means of monthly means), dump those
out to .npy files and plots the anomalies at the end.
If the script does find any previously dumped file, it will load this
file instead of recalculating related variable. Do keep this in mind
while testing... To be on the save side, remove those files before
running the script.